clean up client side test
